refactor: 💡 sort out account book interfaces and function
✅ Closes: #3268 #3263 #3264
