+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.5+134",
+  "version": "0.8.5+135",
   "private": false,
   "scripts": {
     "dev": "next dev",
